# api/openapi.yaml
openapi: 3.0.3
info:
  title: corps (团队) API
  version: 0.6.0
  description: >
    MVP API for the corps multi-tenant collaboration SaaS. This document is the
    single contract for frontend and backend in Phase 3. All paths are prefixed with
    /api/v1/. Every response uses the uniform envelope { code, data, message }.
    Multi-tenant access control: every {wid} route verifies the caller's membership
    row for that workspace, then runs all tenant queries inside a transaction with
    app.workspace_id / app.user_id set (RLS binds at engine level in hardened
    deployments — see db/schema.sql).
    This contract currently defines 57 paths (updated 2026-09-09).
  contact:
    name: MVP Dev Expert Team — Architecture
servers:
  - url: http://localhost:3000
    description: Local development (Next.js Route Handlers)
  - url: https://ghcr.io/{env}/corps
    description: GHCR Docker deployment
    variables:
      env:
        default: staging
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/workspaces/{wid}/tasks/{id}:
    get:
      tags: [tasks]
      summary: Get task detail (incl. assignee/labels/counts)
      responses:
        '200':
          description: Task detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [tasks]
      summary: Update a task (includes drag-to-change-status)
      description: >
        Partial update. Changing `status` is how the kanban drag persists (AC-06).
        Optimistic update on the client; this is the server persistence.
      operationId: updateTask
      parameters:
        - $ref: '#/components/parameters/Wid'
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [tasks]
      summary: Delete a task
      operationId: deleteTask
      parameters:
        - $ref: '#/components/parameters/Wid'
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Null'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/workspaces/{wid}/tasks/batch:
    post:
      tags: [tasks]
      summary: Batch update or delete tasks (max 100 ids)
      description: >
        action=update applies the provided optional fields (status / priority /
        assigneeId / dueDate) to all valid ids; action=delete removes them. Ids not
        belonging to this workspace are reported in `skipped` instead of failing.
        Changing `assigneeId` requires owner/admin (member → 403); the assignee must
        be a member of this workspace (400 otherwise).
      operationId: batchTasks
      parameters:
        - $ref: '#/components/parameters/Wid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids, action]
              properties:
                ids:
                  type: array
                  items: { type: string, format: uuid }
                  minItems: 1
                  maxItems: 100
                action:
                  type: string
                  enum: [update, delete]
                status:
                  type: string
                  enum: [todo, in_progress, review, done]
                priority:
                  type: string
                  enum: [low, medium, high, urgent]
                assigneeId:
                  nullable: true
                  type: string
                  format: uuid
                dueDate:
                  nullable: true
                  type: string
                  format: date-time
      responses:
        '200':
          description: Batch result (partial success)
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 200 }
                  data:
                    type: object
                    required: [updated, deleted, skipped]
                    properties:
                      updated: { type: integer }
                      deleted: { type: integer }
                      skipped:
                        type: array
                        description: Ids not belonging to this workspace.
                        items: { type: string, format: uuid }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ===========================================================================
  # MEMBERS
  # ===========================================================================
...
components:
  securitySchemes:
    cronSecret:
      type: http
      scheme: bearer
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access JWT (15 min) in Authorization header. workspace_id resolved from JWT claim `wid`.
    refreshAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: >
        Better Auth session cookie (httpOnly, secure, 7-day). The /api/v1/auth/refresh
        endpoint reads this cookie to re-issue a workspace-scoped access JWT.
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: HMAC-SHA256 signature of the raw body using the Stripe webhook secret.
  parameters:
    Wid:
      name: wid
      in: path
      required: true
      description: Workspace id. Must equal the JWT `wid` claim; mismatch is 403/404.
      schema: { type: string, format: uuid }
    TaskId:
      name: id
      in: path
      required: true
      description: Task id (UUID).
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Validation error (400)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 400
            data: null
            message: "email must be a valid address"
    Unauthorized:
      description: Missing or invalid token (401)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 401
            data: null
            message: "invalid or expired access token"
    Forbidden:
      description: Authenticated but not permitted (RBAC / tenant mismatch) (403)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 403
            data: null
            message: "role 'member' cannot perform this action"
    NotFound:
      description: Resource unknown to caller (RLS-filtered) (404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 404
            data: null
            message: "workspace not found"
...
    InternalError:
      description: Internal server error (500)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 500
            data: null
            message: "internal server error"
...
  schemas:
...
    Envelope:
      type: object
      required: [code]
      properties:
        code:
          type: integer
          description: "Business status code, mirrors HTTP status code (200=success, 4xx/5xx=error)"
          example: 200
        data:
          nullable: true
        message:
          type: string
          description: "Human-readable status message. Required for error responses (4xx/5xx), optional for success responses (2xx)."
          example: "ok"
    ErrorEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 401
        data:
          nullable: true
          example: null
        message:
          type: string
          example: "invalid or expired access token"
        # L4：events POST 400 等校验失败响应会附带 zod errors 数组，
        # 此前未在 ErrorEnvelope 中文档化。这里补为可选字段（未列入 required），
        # 既覆盖现有带 errors 的 4xx 响应，也不破坏只返回 {code,message,data} 的错误响应。
        errors:
          type: array
          description: >
            Optional field-level validation errors (e.g. zod issues) for 4xx
            responses. Absent on non-validation errors.
          items:
            type: object
            description: A single validation issue (path/message/code as produced by zod).
    Envelope_Null:
      allOf:
        - $ref: '#/components/schemas/Envelope'
      properties:
        data:
          nullable: true
    # L2：Stripe webhook 200 实际返回 { received: true } 或 { received: true, duplicate: true }，
    # 此前误用 Envelope_Null（data:null），与实现不一致。这里定义专用 envelope，
    # data 形状为 { received: boolean, duplicate?: boolean }，与 route.ts 两处 200 应答对齐。
...
    Envelope_Task:
      allOf:
        - $ref: '#/components/schemas/Envelope'
      properties:
        data:
          $ref: '#/components/schemas/Task'
...
    Task:
      type: object
      required: [id, workspaceId, title, status, priority, sortOrder, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigneeId:
          nullable: true
          type: string
          format: uuid
        dueDate:
          nullable: true
          type: string
          format: date-time
        sortOrder:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateTaskRequest:
      type: object
      description: Partial update. Any subset of fields may be sent.
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 5000
        status:
          type: string
          enum: [todo, in_progress, review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigneeId:
          nullable: true
          type: string
          format: uuid
        dueDate:
          nullable: true
          type: string
          format: date-time
        sortOrder:
          type: number
          format: double
        blocked:
          type: boolean
          description: Whether the task is blocked.
        blockedReason:
          type: string
          maxLength: 500
          nullable: true
          description: Optional reason text when blocked is true.
        shareToken:
          type: string
          maxLength: 64
          nullable: true
          description: >-
            Public share token. Send "rotate" to generate a new token, null to cancel sharing.

    # ---- members (shape mirrors the actual listMembers response) ----
...